scripts/requirements.txt
autopep8==2.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pycodestyle==2.11.1
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1